Separate headers for mapping and index parameters.
